% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{file_name_2_sample_name}
\alias{file_name_2_sample_name}
\title{file_name_2_sample_name(file_name)}
\usage{
file_name_2_sample_name(file_name)
}
\arguments{
\item{file_name}{file full path}
}
\value{
sample_name extracted sample name
}
\description{
extract sample name from file full path, removing directory
             structure and file extension
}
\examples{
file_name_2_sample_name("../ime_uzorka.dat")
}
\author{
K. Juraić
}
