% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{gixrd_average_lst}
\alias{gixrd_average_lst}
\title{gixrd_read_file_lst(file_names, save_to_file = TRUE)}
\usage{
gixrd_average_lst(file_names, save_to_file = FALSE)
}
\arguments{
\item{file_names}{array of file names with GIXRD data (full paths)}

\item{save_to_file}{Should be the averaged data saved to file}
}
\value{
list of data frames with averaged data
}
\description{
Automaticaly read multiple files with GIXRD data and
             average them and write averaged data to files
}
\examples{
\dontrun{gixrd_aveeage_lst(fnms)}
}
