% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_mcx_xrd}
\alias{read_mcx_xrd}
\title{read GIXRD data from MCX file}
\usage{
read_mcx_xrd(xrd_file = tcltk::tk_choose.files(multi = FALSE))
}
\arguments{
\item{xrd_file}{name of filme with MCX GIXRD data}
}
\value{
list(xrd_file, alpha_i, data.frame(tth, intensity, background))
}
\description{
read GIXRD data from MCX file
}
\examples{
\dontrun{read_mcx_xrd()}
}
