% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_mcx_xrd_multi}
\alias{read_mcx_xrd_multi}
\title{read multiple MCX XRD files}
\usage{
read_mcx_xrd_multi(xrd_files = tcltk::tk_choose.files())
}
\arguments{
\item{xrd_files}{array with full file namse}
}
\value{
list of xrd data (for each file name single list element)
}
\description{
read multiple MCX XRD files
}
\examples{
\dontrun{read_mcx_xrd_multi}
}
