% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_mercury_hkl}
\alias{read_mercury_hkl}
\title{read_mercury_hkl(file_name)}
\usage{
read_mercury_hkl(file_name = file.choose())
}
\arguments{
\item{file_name}{mercurt *.hkl file name}
}
\value{
data.frame(h, k, l, d, f2, mult)
}
\description{
read mercury hkl file with d spacing [Angstrom],
             structure factor and multiplicit for (hkl) difraction peaks
}
\examples{
\dontrun{read_mercury_hkl("ZnO.hkl")}

}
\author{
K. Juraic
}
