% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_spec}
\alias{read_spec}
\title{Read GIXRD data from file (Buljan instrument)}
\usage{
read_spec(file_name)
}
\arguments{
\item{file_name}{GIXRD data file name}
}
\value{
list with GIXRD data
}
\description{
Read GIXRD data from file (Buljan instrument) to list.
             In one file are writen all scans for one single sample.
             Each list element coresponds to one scan in file.
             Columns description for file format: MUST BE ADDED
}
\examples{
            \dontrun{read_spec("gixrd_name")}
}
\author{
K. Juraic
}
