% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_xy}
\alias{read_xy}
\title{read XRD data from xy file}
\usage{
read_xy(file_name)
}
\arguments{
\item{file_name}{XRD data filename (full path)}
}
\value{
data.frame(tth, intensity)
}
\description{
read XRD data from xy file. Can be generated from raw instrument
             data with powDLL software. Data are stored as two
             column (tth, intensity).
}
\examples{
\dontrun{read_xy(""XRD_pattern.xy)}

}
\author{
K. Juraić
}
