% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice_parameters.R
\name{tetragonal_d_2_c}
\alias{tetragonal_d_2_c}
\title{tetragonal_d_2_c(h, k, l, d, a)}
\usage{
tetragonal_d_2_c(h, k, l, d, a)
}
\arguments{
\item{h}{Muller index}

\item{k}{Muller index}

\item{l}{Muller index}

\item{d}{d-spacing}

\item{a}{lattice paramerr}
}
\value{
c lattice parameter of hexagonal crystall structure
}
\description{
calculate lattice parameter c for tetragona crystall structure
             from d-spacing from known Muller indexes (h,k,l) and lattice
             paramter a. For specioal case when Muller index h = 0, c can be
             calculated independently on Muller index a
}
\examples{
tetragonal_d_2_c(0, 0, 1, 1.59, 1)
}
\author{
K. Juraić
}
