% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice_parameters.R
\name{tetragonal_hkl_2_d}
\alias{tetragonal_hkl_2_d}
\title{tetragonal_hkl_2_d(h, k, l, a, c)}
\usage{
tetragonal_hkl_2_d(h, k, l, a, c)
}
\arguments{
\item{h}{millerov indeks}

\item{k}{millerov indeks}

\item{l}{millerov index}

\item{a}{lattice parameter}

\item{c}{lattice parameter}
}
\value{
d-spacing : razmak mrežnih ravnina
}
\description{
For tetragona crystall structure calculate d-spacing by using
             known lattice paramteres (a, c) and Muller indexes
}
\examples{
\dontrun{tetragonal_hkl_2_d(h = 1, k = 1, l = 0
                                      a = 4.73727, c = 3.186383)}
}
\author{
K. Juraic
}
