% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice_parameters.R
\name{tth_2_d}
\alias{tth_2_d}
\title{d_tth_2_d(tth, lambda)}
\usage{
tth_2_d(tth, lambda = 1.54056)
}
\arguments{
\item{tth}{2*theta position of XRD peak in degrees}

\item{lambda}{X-ray wavelength in angstroms or nanometers}
}
\value{
d-spacing in same units as wavelength
}
\description{
calculate d-spacing for tth position of XRD peak in 2*theta scale
             tth should be given in degrees
}
\examples{
tth_2_d(28.1, 1.54056)
}
\author{
K. Juraić
}
