% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gixrd_ggplot.R
\name{xrd_ggplot_add_peak_pos}
\alias{xrd_ggplot_add_peak_pos}
\title{add marks for XRD peak positions (tth) to gglot XRD graph}
\usage{
xrd_ggplot_add_peak_pos(p, peaks_df, stick_y = c(-100, 0))
}
\arguments{
\item{p}{ggplot XRD graph}

\item{peaks_df}{data.frame with column tth (xrd peaks possitions)}

\item{stick_y}{(y_min, y(max)) y coordinates of stick stard and end}
}
\description{
add marks for XRD peak positions (tth) to gglot XRD graph
}
\examples{
\dontrun{xrd_ggplot_add_peak_pos(p, df, c(0,100))}
}
\author{
K. Juraic
}
