% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gixrd_ggplot.R
\name{xrd_ggplot_style}
\alias{xrd_ggplot_style}
\title{add GIXRD style to ggplot of experimental data}
\usage{
xrd_ggplot_style(p)
}
\arguments{
\item{p}{ggplot2 with experimental data}
}
\value{

}
\description{
add GIXRD style to ggplot of experimental data
}
\examples{
\dontrun{xrd_ggplot_style(p)}
}
\author{
K. Juraic
}
