% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{xrd_intensity_shift}
\alias{xrd_intensity_shift}
\title{Shift intensity in steps for presenting data ain shifted plot}
\usage{
xrd_intensity_shift(dat_df, shift_factor = 0)
}
\arguments{
\item{dat_df}{data.frame(tth, intensity, name) with XRD data}

\item{shift_factor}{single number or array with values to shift intensity data}
}
\value{
data.frame(tth, intensity, name)
}
\description{
Shift of intensity data for shifted plot. If shift factor is
             single number it is constructed intensity shift array with
             equidistant shift factor for each diffractogram in data.frame.
             Array lengt is equial to number of diffractograms in data frames.
             If shift array is provided as aaray of same length as number of
             diffractograms in data.frames is directly uses to shift data in
             data.frames.
}
\examples{
\dontrun{xrd_intensity_shift(dat_df= xrd_data, shift_factor = 100)}
}
\author{
K. Juraić
}
