% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_xrd_data}
\alias{read_xrd_data}
\title{read_xrd_data(data_dir =  tcltk::tk_choose.dir(), pattern = "")}
\usage{
read_xrd_data(data_dir = tcltk::tk_choose.dir(), pattern = "")
}
\arguments{
\item{data_dir}{folder path with xrd files}

\item{pattern}{file extension (.xy, .raw)}
}
\value{
data.frame(tth, intesity)
}
\description{
read xrd experimental data from folder selected by file extension
             At the moment applicable for .xy and Bruker .raw format
}
\author{
K. Juraić
}
